better join and quit raid

+++ Raid/src/main/java/com/scatena/marcello/Raid/controller/raidController.java
@@ -8,6 +8,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -42,7 +43,7 @@ public class raidController {
 		List<RaidDTO> raidDTOList = new ArrayList();
 		raidsList.forEach(raid -> {
 			PokemonDTO pokemon = raidService.getPokemon(raid.getPokemonNumber());
-			RaidDTO raidDTO = new RaidDTO(raid.getId(), raid.getHostFC(), pokemon);
+			RaidDTO raidDTO = new RaidDTO(raid.getId(), raid.getHostFC(), raid.getGuestFC1(),raid.getGuestFC2(),raid.getGuestFC3(), pokemon);
 			raidDTOList.add(raidDTO);
 			});
 		return raidDTOList;
@@ -62,31 +63,77 @@ public class raidController {
 	
 	@CrossOrigin
 	@PostMapping
-	public void createRaid(@RequestBody Raid raid) {
+	public Long createRaid(@RequestBody Raid raid) {
 		raidRepository.save(raid);
+		raidRepository.flush();
+		return raid.getId();
+		
 	}
 	
 	@CrossOrigin
-	@PutMapping("/{id}/{guestFC}")
+	@PutMapping("/join/{id}/{guestFC}")
 	@Transactional
-	public Raid joinRaid(@PathVariable Long id, @PathVariable Long guestFC) {
+	public ResponseEntity<Raid> joinRaid(@PathVariable Long id, @PathVariable Long guestFC) {
 		Optional<Raid> optRaid = raidRepository.findById(id);
 		Raid raid = optRaid.get();
-		if(raid.getGuestFC1()==null)
+		if(raid.getGuestFC1()==null) {
 			raid.setGuestFC1(guestFC);
-		else if(raid.getGuestFC2()==null)
+			return ResponseEntity.ok(raid);
+		}
+			
+		else if(raid.getGuestFC2()==null) {
 			raid.setGuestFC2(guestFC);
-		else if(raid.getGuestFC3()==null)
+			return ResponseEntity.ok(raid);
+		}
+			
+		else if(raid.getGuestFC3()==null) {
 			raid.setGuestFC3(guestFC);
-		return raid;
+			return ResponseEntity.ok(raid);
+		}
+		else {
+			return ResponseEntity.badRequest().build();
+		}	
+		
+	}
+	
+	@CrossOrigin
+	@PutMapping("/quit/{id}/{guestFC}")
+	@Transactional
+	public ResponseEntity<Raid> quitRaid(@PathVariable Long id, @PathVariable Long guestFC) {
+		Optional<Raid> optRaid = raidRepository.findById(id);
+		Raid raid = optRaid.get();
+		if(raid.getGuestFC1()==guestFC) {
+			raid.setGuestFC1(null);
+			return ResponseEntity.ok(raid);
+		}
+			
+		else if(raid.getGuestFC2()==guestFC) {
+			raid.setGuestFC2(null);
+			return ResponseEntity.ok(raid);
+		}
+			
+		else if(raid.getGuestFC3()==guestFC) {
+			raid.setGuestFC3((long) 344);
+			return ResponseEntity.ok(raid);
+		}
+		else {
+			return ResponseEntity.badRequest().build();
+		}	
+		
 	}
 	
 	@CrossOrigin
 	@DeleteMapping("/{id}")
 	@Transactional
-	public void remover(@PathVariable Long id) {
-
-		raidRepository.deleteById(id);
+	public ResponseEntity<Raid> deleteRaid(@PathVariable Long id) {
+		Optional<Raid> optional = raidRepository.findById(id);
+		if (optional.isPresent()) {
+			raidRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		
+		return ResponseEntity.notFound().build();
+		
 		
 	}
 }
